+++ src/main/java/com/github/msemys/esjc/StreamPosition.java
@@ -8,12 +8,12 @@ public class StreamPosition {
     /**
      * The first event in a stream.
      */
-    public static final int START = 0;
+    public static final long START = 0;
 
     /**
      * The last event in the stream.
      */
-    public static final int END = -1;
+    public static final long END = -1;
 
     private StreamPosition() {
     }
