+++ src/main/java/com/github/msemys/esjc/operation/ReadStreamEventsBackwardOperation.java
@@ -16,14 +16,14 @@ import static com.github.msemys.esjc.util.Strings.defaultIfEmpty;
 public class ReadStreamEventsBackwardOperation extends AbstractOperation<StreamEventsSlice, ReadStreamEventsCompleted> {
 
     private final String stream;
-    private final int fromEventNumber;
+    private final long fromEventNumber;
     private final int maxCount;
     private final boolean resolveLinkTos;
     private final boolean requireMaster;
 
     public ReadStreamEventsBackwardOperation(CompletableFuture<StreamEventsSlice> result,
                                              String stream,
-                                             int fromEventNumber,
+                                             long fromEventNumber,
                                              int maxCount,
                                              boolean resolveLinkTos,
                                              boolean requireMaster,
